+++ test/test_join.c
@@ -42,7 +42,7 @@
 #include "os.h"
 #include "os_test.h"
 
-#define TEST_TASKS ((size_t)10)
+#define TEST_TASKS ((unsigned)10)
 #define TEST_CYCLES ((os_atomic_t)1000)
 
 static os_task_t task_worker1;
